refactor(entity): 将 Comment 类中的 Long 类型字段改为 Integer 类型
- 将 id、postId 和 userId 字段的类型从 Long 改为 Integer
- 这一变更可以减少存储空间的使用，并提高数据处理效率

+++ backend/src/main/java/cn/hmck/entity/Comment.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @TableName("er_comment")
 public class Comment implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 评论ID
+    private Integer id;  // 评论ID
 
-    private Long postId;  // 动态ID
+    private Integer postId;  // 动态ID
 
-    private Long userId;  // 发布评论的用户ID
+    private Integer userId;  // 发布评论的用户ID
 
     private Integer type;  // 评论类型，1为评论，2为点赞
